==> 09_symbionts_NS/01_combine_fungal_guilds.ipynb <==
import pandas as pd
import numpy as np
from sklearn import preprocessing
import matplotlib.pyplot as plt
import seaborn as sns

# DATA

dmeta = pd.read_csv("../01_data_prep/dataPrep_01_overview_ITS_META.tsv", sep="\t", header=0)
dmeta['SampleID'] = dmeta['long.name'].apply(lambda x: x.replace('.','-').replace("X",""))
print(dmeta.shape)
dmeta.head()

samples = dmeta["SampleID"].drop_duplicates().values
print(len(samples))
#samples

df = pd.read_csv("/mnt/d/NRCan/2022_MicrobiomeQuebecMexico/DATA/ITS/Full/Peuplier_REPRISE_ITS_oct2022.ASV_table_norarefaction_dnNA.guilds.txt", sep='\t', header=0)
df.head()

## Abundance and taxonomy tables 

#columns_its = [ele for ele in df.columns if 'ITS' in ele]
columns_its = samples
columns_tax = [ele for ele in df.columns if 'ITS' not in ele]
df_abu = df[columns_its]
df_abu = df_abu.set_index(df['#OTU ID'])
print(df_abu.shape)
df_tax = df[columns_tax]
df_tax = df_tax.set_index('#OTU ID')
print(df_tax.shape)
df_tax.head()

df_tax['group'] = df_tax['Guild'].fillna('Other')
df_tax['group'] = df_tax['group'].apply(lambda x: x if x != "-" else "Other")
df_tax['group'] = df_tax.apply(lambda x: x['group'] if x['Confidence Ranking'] in ["Probable","Highly Probable"] else "Other", axis=1)
df_tax['group'] = df_tax['group'].apply(lambda x: 'Mixed' if '-' in str(x) else str(x))
df_tax.head()

df_guilds = df_tax[['group']]
df_guilds

## Subset soil samples

columns_soil = [ele for ele in df_abu if 'Soil' in ele]
df_abu_soil = df_abu[columns_soil]
print(df_abu_soil.shape)
df_abu_soil.head()

# Sum abundances across guilds

df_guilds.head()

print(df_guilds.shape)
print(df_abu_soil.shape)
df_abu_soil.head()

df_sum = df_abu_soil.groupby(df_guilds['group']).sum()
df_sum.to_csv("guilds_01_read_stands.csv", sep=",", header=True, index=True)

df_sum.head()

dmeta.head()

dt_sum = df_sum.T
stack_guilds = dt_sum.stack().reset_index().rename(columns={'level_0' : 'sample', 0 : 'count'})
df_meta_summed = pd.merge(stack_guilds, dmeta, left_on = 'sample', right_on = 'SampleID', how='left')
df_meta_summed

df_meta_summed.to_csv("guilds_01_read_stands_meta.csv", sep=",", header=True, index=False)
