==> 09_symbionts_NS/07_combine_nitrogen_fixers.ipynb <==
import pandas as pd
import numpy as np
from sklearn import preprocessing
import matplotlib.pyplot as plt
import seaborn as sns

# DATA

dmeta = pd.read_csv("../01_data_prep/dataPrep_01_overview_16s_META.tsv", sep="\t", header=0)
dmeta['SampleID'] = dmeta['long.name'].apply(lambda x: x.replace('.','-').replace("X",""))
print(dmeta.shape)
dmeta.head()

samples = dmeta["SampleID"].drop_duplicates().values
print(len(samples))
print(len([ele for ele in samples if 'Soil' in ele]))
dmeta.head()

df = pd.read_csv("/mnt/d/NRCan/2022_MicrobiomeQuebecMexico/DATA/16S/Full/Peuplier_REPRISE_16S_oct2022.ASV_table_norarefaction_dnNA.tsv", sep='\t', header=0)
df.head()

## Abundance table

columns_tab = samples
df_abu = df[columns_tab]
df_abu = df_abu.set_index(df['#OTU ID'])
print(df_abu.shape)
df_abu.head()

# Taxonomy table

dtax = pd.read_csv("../../2022_MicrobiomeAspenGreenhouse/01_data_prep/dataPrep_01_overview_16S_TAX.tsv", sep="\t", header=0)
dtax["Phy"] = dtax["Phylum"].fillna('Unknown')
dtax["Fam"] = dtax["Family"].fillna('Unknown')
dtax["Fam"] = dtax["Fam"].apply(lambda x: x.replace("f__",""))
dtax["Gen"] = dtax["Genus"].fillna('Unknown')
dtax["Gen"] = dtax["Gen"].apply(lambda x: x.replace("g__",""))
dtax["Spe"] = dtax["Species"].fillna('Unknown')
dtax["Spe"] = dtax["Spe"].apply(lambda x: x.replace("s__",""))
dtax.head()

# Selection of nitrogen-fixing bacteria
nf1 = dtax[dtax['Gen'].str.contains('Frankia')].index.values.tolist()
nf2 = dtax[dtax['Phy'].str.contains('Cyanobacteria')].index.values.tolist()
nf3 = dtax[dtax['Gen'].str.contains('rhizobium')].index.values.tolist()
nf4 = dtax[dtax['Gen'].str.contains('Rhizobium')].index.values.tolist()
nf5 = dtax[dtax['Gen'].str.contains('Ensifer|Azotobacter|Azospirillum|Beijerinckia|Azoarcus', case=False, na=False)].index.values.tolist()
print(len(nf1), len(nf2), len(nf3), len(nf4), len(nf5))
nfs = list(set(nf1 + nf2 + nf3 + nf4 + nf5))
print(len(nfs))

print(len(nf3), len(nf4))

dtax['asv'] = dtax.index
dtax['NF'] = dtax['asv'].apply(lambda x: 'NF' if x in nfs else 'nonNF')
df_guilds = dtax[['NF']]
df_guilds

pd.DataFrame({"ASV": dtax[dtax['NF'] == 'NF'].index.values}).to_csv("nitro_02_read_NF_ASVs.txt", sep="\t", header=True, index=False)

## Subset soil samples

columns_soil = [ele for ele in df_abu if 'Soil' in ele]
df_abu_soil = df_abu[columns_soil]
print(df_abu_soil.shape)
df_abu_soil.head()

# Sum abundances across guilds

df_guilds.head()

print(df_guilds.shape)
print(df_guilds[['NF']].drop_duplicates().shape)
print(df_abu_soil.shape)
df_abu_soil.head()

df_sum = df_abu_soil.groupby(df_guilds['NF']).sum()
df_sum.to_csv("nitro_02_read.csv", sep=",", header=True, index=True)
df_sum.head()

dmeta.head()

dt_sum = df_sum.T
stack_guilds = dt_sum.stack().reset_index().rename(columns={'level_0' : 'sample', 0 : 'count'})
df_meta_summed = pd.merge(stack_guilds, dmeta, left_on = 'sample', right_on = 'SampleID', how='left')
df_meta_summed

df_meta_summed.to_csv("nitro_02_read_meta.csv", sep=",", header=True, index=False)
